fixed more bugs when on mobile, centered testimonial cards, hid world model on mobile
