+++ client/src/App.jsx
@@ -1,21 +1,48 @@
 import { BrowserRouter } from "react-router-dom";
+import React, { useEffect, useState } from "react";
 
 import { About, Contact, Experience, Feedbacks, Hero, Navbar, Tech, Works, StarsCanvas} from './components';
 
 const App = () => {
+
+  const [isMobile, setIsMobile] = useState(false);
+
+  useEffect(() => {
+    // Add a listener for changes to the screen size
+    const mediaQuery = window.matchMedia("(max-width: 900px)");
+
+    // Set the initial value of the `isMobile` state variable
+    setIsMobile(mediaQuery.matches);
+
+    // Define a callback function to handle changes to the media query
+    const handleMediaQueryChange = (event) => {
+      setIsMobile(event.matches);
+    };
+
+    // Add the callback function as a listener for changes to the media query
+    mediaQuery.addEventListener("change", handleMediaQueryChange);
+
+    // Remove the listener when the component is unmounted
+    return () => {
+      mediaQuery.removeEventListener("change", handleMediaQueryChange);
+    };
+  }, []);
+
+
+
   return (
     <BrowserRouter>
       <div className='relative z-0 bg-primary'>
         <div className='bg-hero-pattern bg-cover bg-no-repeat bg-center'>
           <Navbar/>
-          <Hero/>
+          <Hero isMobile={isMobile}/>
         </div>
         <About/>
         <Experience/>
         <Works/>
         <Feedbacks/>
         <div className="relative z-0">
-          <Contact/>
+          <Contact isMobile={isMobile}/>
           <StarsCanvas/>
         </div>
       </div>
