+++ client/src/components/Feedbacks.jsx
@@ -53,7 +53,7 @@ const Feedbacks = () => {
         </motion.div>
       </div>
 
-      <div className={`${styles.paddingX} -mt-20 pb-14 flex flex-wrap gap-7`}>
+      <div className={`${styles.paddingX} -mt-20 pb-14 flex flex-wrap gap-7 justify-center`}>
         {testimonials.map((testimonial, index) => (
           <FeedbackCard
           key={testimonial.name}
